(YE-1) Added UI skeleton for the event cards
The event card itself contains elements that may be removed in the
future such as the registration button. That could be changed to a
redirection to the source of the information, such as a discord channel
or instagram post.
Need to fix for api calls, as only placeholder information is currently
used.
    - Such as removing the static information with remote sources
Improved responsiveness for different screen widths.
    - May want to fix overlapping of filter grid fields
Fixed minor layering issues with hover/dropdown menus